Fixed error related issues. Event system prototype.
