+++ src/lib/events/src/lib.rs
@@ -2,3 +2,4 @@
 #[cfg(test)]
 mod tests;
 pub mod infrastructure;
+pub mod errors;
