+++ src/lib/net/encryption/src/errors.rs
@@ -1,7 +1,7 @@
 use thiserror::Error;
 
 #[derive(Debug, Clone, Error)]
-pub enum EventsError {
+pub enum NetEncryptionError {
     #[error("Something failed lol")]
     SomeError,
 }
