--- src/world/src/errors.rs
@@ -1,7 +1,0 @@
-use thiserror::Error;
-
-#[derive(Debug, Clone, Error)]
-pub enum WorldError {
-    #[error("Something failed lol")]
-    SomeError,
-}
